Changed the log4j dependency from log4j 1.x to log4j 2.x. - This change had the effect, that all classes that used the log4j had to be adapted. - The GUI class used in addition PatternLayout and ConsoleAppender, which changed in log4j 2. - The update commented it out, since the code is not used at any use cases in the moment. When the powerflowsim is back in production, the logging should be checked again.
