--- core/powerflow/dependency-reduced-pom.xml
@@ -1,133 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.fortiss</groupId>
-  <artifactId>powerflow</artifactId>
-  <name>powerflow</name>
-  <version>0.1</version>
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.1</version>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-      <plugin>
-        <artifactId>maven-shade-plugin</artifactId>
-        <version>3.2.1</version>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <transformers>
-            <transformer />
-          </transformers>
-          <filters>
-            <filter>
-              <artifact>*:*</artifact>
-              <excludes>
-                <exclude>META-INF/*.SF</exclude>
-                <exclude>META-INF/*.DSA</exclude>
-                <exclude>META-INF/*.RSA</exclude>
-              </excludes>
-            </filter>
-          </filters>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.10</version>
-      <scope>test</scope>
-      <exclusions>
-        <exclusion>
-          <artifactId>hamcrest-core</artifactId>
-          <groupId>org.hamcrest</groupId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-  </dependencies>
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.swinglabs</groupId>
-        <artifactId>swing-layout</artifactId>
-        <version>1.0.3</version>
-      </dependency>
-      <dependency>
-        <groupId>log4j</groupId>
-        <artifactId>log4j</artifactId>
-        <version>1.2.17</version>
-      </dependency>
-      <dependency>
-        <groupId>com.github.rwl</groupId>
-        <artifactId>jpower</artifactId>
-        <version>4.0.0</version>
-      </dependency>
-      <dependency>
-        <groupId>org.apache.commons</groupId>
-        <artifactId>commons-lang3</artifactId>
-        <version>3.1</version>
-      </dependency>
-      <dependency>
-        <groupId>jgraph</groupId>
-        <artifactId>jgraph</artifactId>
-        <version>5.13.0.0</version>
-      </dependency>
-      <dependency>
-        <groupId>org.tinyjee.jgraphx</groupId>
-        <artifactId>jgraphx</artifactId>
-        <version>2.3.0.5</version>
-      </dependency>
-      <dependency>
-        <groupId>junit</groupId>
-        <artifactId>junit</artifactId>
-        <version>4.10</version>
-      </dependency>
-      <dependency>
-        <groupId>org.fortiss</groupId>
-        <artifactId>cim15</artifactId>
-        <version>1.0.0</version>
-      </dependency>
-      <dependency>
-        <groupId>net.sourceforge.parallelcolt</groupId>
-        <artifactId>parallelcolt</artifactId>
-        <version>0.10.1</version>
-      </dependency>
-      <dependency>
-        <groupId>com.google.api.client</groupId>
-        <artifactId>google-api-client-repackaged-com-google-common-base</artifactId>
-        <version>1.2.3-alpha</version>
-      </dependency>
-      <dependency>
-        <groupId>com.google.code.google-collections</groupId>
-        <artifactId>google-collect</artifactId>
-        <version>snapshot-20080530</version>
-      </dependency>
-      <dependency>
-        <groupId>swt</groupId>
-        <artifactId>forms</artifactId>
-        <version>3.0m8</version>
-      </dependency>
-      <dependency>
-        <groupId>com.github.rwl</groupId>
-        <artifactId>jplasma</artifactId>
-        <version>1.2.0</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-</project>
