+++ core/powerflow/src/main/java/org/fortiss/powerflowsim/simulation/internal/BusBranchModel.java
@@ -15,8 +15,10 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import org.apache.log4j.Level;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.core.config.Configurator;
 import org.fortiss.powerflowsim.model.AdditionalInformationMapping;
 import org.fortiss.powerflowsim.model.CimModel;
 import org.fortiss.powerflowsim.model.ExtendedTopologyInformation;
@@ -44,7 +46,7 @@ import CIM15.IEC61970.Wires.PowerTransformer;
  */
 public class BusBranchModel {
 	
-	private static Logger log = Logger.getLogger(BusBranchModel.class);
+	private static Logger log = LogManager.getLogger(BusBranchModel.class);
 	
 	//public static ActorTopology actorTopology = null;
 	
@@ -88,7 +90,9 @@ public class BusBranchModel {
 	 */
 	public BusBranchModel(CimModel model) {
 		
-		log.setLevel(Level.WARN);
+		// previous solution only as reminder 
+		// log.setLevel(Level.WARN);
+		Configurator.setLevel(log.getName(), Level.WARN);
 
 		// get the number of busses (represented through a topological node in CIM) and lines (represented through a ACLineSegment CIM object)
 		numberOfNodes = model.getTopologicalNodes().size();
