+++ core/powerflow/src/test/java/org/fortiss/powerflowsim/test/GridLabDTest.java
@@ -11,10 +11,10 @@ package org.fortiss.powerflowsim.test;
 
 import java.io.File;
 
-import org.apache.log4j.ConsoleAppender;
-import org.apache.log4j.Level;
-import org.apache.log4j.Logger;
-import org.apache.log4j.PatternLayout;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.core.config.Configurator;
 import org.fortiss.powerflowsim.exporters.GLMExporter;
 import org.fortiss.powerflowsim.importers.SampleCases;
 import org.fortiss.powerflowsim.model.CimModel;
@@ -24,16 +24,21 @@ import org.junit.Before;
 import org.junit.Test;
 
 public class GridLabDTest {
-	private static Logger log = Logger.getRootLogger();
+	private static Logger log = LogManager.getRootLogger();
 	private CimModel model9bus;
 
 	@Before
 	public void createLogger() {
-		PatternLayout layout = new PatternLayout("%d{ISO8601} %-5p [%t] %c: %m%n");
-		ConsoleAppender consoleAppender = new ConsoleAppender(layout);
-		log.addAppender(consoleAppender);
+		//PatternLayout layout = new PatternLayout("%d{ISO8601} %-5p [%t] %c: %m%n");
+		//ConsoleAppender consoleAppender = new ConsoleAppender(layout);
+		//log.addAppender(consoleAppender);
+		
 		// ALL | DEBUG | INFO | WARN | ERROR | FATAL | OFF:
-		log.setLevel(Level.ALL);
+		//log.setLevel(Level.ALL);
+		
+		// previous solution only as reminder 
+		// log.setLevel(Level.WARN);
+		Configurator.setLevel(log.getName(), Level.ALL);
 	}
 
 	@Before
