+++ core/powerflow/src/test/java/org/fortiss/powerflowsim/test/SolverTest.java
@@ -11,10 +11,10 @@ package org.fortiss.powerflowsim.test;
 
 import static org.junit.Assert.assertTrue;
 
-import org.apache.log4j.ConsoleAppender;
-import org.apache.log4j.Level;
-import org.apache.log4j.Logger;
-import org.apache.log4j.PatternLayout;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.core.config.Configurator;
 import org.fortiss.powerflowsim.importers.SampleCases;
 import org.fortiss.powerflowsim.importers.TopologyGenerator;
 import org.fortiss.powerflowsim.model.CimModel;
@@ -26,17 +26,21 @@ import org.junit.Before;
 import org.junit.Test;
 
 public class SolverTest {
-	private static Logger log = Logger.getRootLogger();
+	private static Logger log = LogManager.getRootLogger();
 	private CimModel model9bus;
 	private CimModel model1000bus;
 
 	@Before
 	public void createLogger() {
-		PatternLayout layout = new PatternLayout("%d{ISO8601} %-5p [%t] %c: %m%n");
-		ConsoleAppender consoleAppender = new ConsoleAppender(layout);
-		log.addAppender(consoleAppender);
+		//PatternLayout layout = new PatternLayout("%d{ISO8601} %-5p [%t] %c: %m%n");
+		//ConsoleAppender consoleAppender = new ConsoleAppender(layout);
+		//log.addAppender(consoleAppender);
 		// ALL | DEBUG | INFO | WARN | ERROR | FATAL | OFF:
-		log.setLevel(Level.ALL);
+		//log.setLevel(Level.ALL);
+		
+		// previous solution only as reminder 
+		// log.setLevel(Level.WARN);
+		Configurator.setLevel(log.getName(), Level.ALL);
 	}
 
 	@Before
